move validateBase64EncodedUuid function
