+++ webapi/src/main/scala/org/knora/webapi/messages/v2/responder/resourcemessages/ResourceMessagesV2.scala
@@ -704,8 +704,6 @@ object CreateResourceRequestV2 {
                  throw BadRequestException(s"<$iri> is not a Knora resource IRI")
                }
 
-               stringFormatter.validateUUIDOfResourceIRI(iri)
-
                if (!iri.getProjectCode.contains(projectInfoResponse.project.shortcode)) {
                  throw BadRequestException(s"The provided resource IRI does not contain the correct project code")
                }
@@ -907,7 +905,6 @@ object UpdateResourceMetadataRequestV2 {
       resourceIri <- obj.getRequiredIdValueAsKnoraDataIri
                        .filterOrElseWith(_.isKnoraResourceIri)(it => ZIO.fail(s"Invalid resource IRI: <$it>"))
                        .mapError(BadRequestException(_))
-      _ <- ZIO.serviceWithZIO[StringFormatter](sf => ZIO.attempt(sf.validateUUIDOfResourceIRI(resourceIri)))
     } yield resourceIri
 
   private def getResourceClassIri(body: JsonLDObject) =
